change mm ft in to int

+++ sql_scripts/create_iso_container_size_code_1.sql
@@ -6,8 +6,8 @@ GO
 
 CREATE TABLE ISOSizeCode1
 (
-    [Container length (mm)] decimal,
-    [Container length (ft)] decimal,
-    [Container length (in)] decimal,
+    [Container length (mm)] int,
+    [Container length (ft)] int,
+    [Container length (in)] int,
     [Code character]        nvarchar(255)
 );
